Make pyScss compatible again with python 2.5.
The change introduced in changeset 4da61e4 broke the compilataion of scss files
with Python 2.5 : the parameter 'break_on_hyphens' of  textwrap.TextWrapper()
constructor has been added in Python 2.6.

However, the readme claims as requirement "python >= 2.5".

Another solution would be to remove completely support for python 2.5.

On a sidenote: I maintain (when I have some time) to maintain the
extension Flask-Scss that uses pyScss as its compilation engine,
and as Flask supports python 2.5, I am required to support it too, which
motivates this request.