Remove _skip_word_re.
This was an optimization to short-circuit evaluating any expressions
that don't appear to contain Sass syntax, but it was wrong: the
particular case I ran across was `foo, bar` not being parsed as a list!

Removing it added only 3% to the test suite runtime.  We can see about
resurrecting the idea later, after the suite actually passes.  :)