Remove the "thin ice" logic from the block locators.
It took some experimentation to figure out what this actually does.  It
parses "a,\nb\nc{...}" such that "a,\nb" is considered a separate block
or property or whatever from "c{...}".  I think this is left over from
how xCSS treated blocks (with regexes!), but it looks outright wrong to
me, since a CSS parser would treat that as equivalent to "a, b c {...}".

So away it goes!