Fix the fallback behavior of / and - on non-strings.
The default is to treat these like a single unit, so "a" - "b" produces
"a"-"b".  The code was neglecting to use String.unquoted, so the
resulting string would have even more quotes around it, which was...
incorrect.

This mostly had an effect on using CSS syntax like top/1px inside a
variable, where division is evaluated.  The result was quoted, which
changed the semantics.