Put @imported rules in the place they were imported.  Fixes #352
There was some old code that grouped rules by the literal file they came
from -- even if @imported -- solely for the sake of some debug code that
I'm not convinced is particularly useful.  I removed the grouping, which
ensures that an @import in the middle of a file will correctly leave its
rules in that position.

Also inlined a few very short methods that weren't used anywhere else
and mostly just made browsing the code more confusing than necessary.  I
think this might break the repl, but it was already broken, so.