Fix DeprecationWarning on inspect.getargspec in python3
inspect.getargspec() is deprecated in python3 and the usage of
inspect.signature() or inspect.getfullargspec() is recommended.
The only different between getargspec and getfullargspec is
that "keywords" attribute in getargspec() return value corresponds
to "varkw" attribute in case of getfullargspec().
This commit changes the code to try getfullargspec() first and
then fallback to getargspec().